+++ src/components/AccountList.tsx
@@ -11,7 +11,7 @@ import {useProject} from "providers/Project/projectHooks";
 import Avatar from "components/Avatar";
 import styled from "@emotion/styled";
 import {ExportButton} from "components/ExportButton";
-import {getParams, isUUUID} from "../util/uuid";
+import {getParams, isUUUID} from "../util/url";
 
 function getDeployedContracts(account: Account): string {
   const contracts = account.deployedContracts.map(
